Make use of `strtonum()` more consistent
Explicitly downcast to int.  Style the typecast consistently.
Consistently name the buffer into which `strtonum()` writes its error
message `errstr`.  Consistently explicitly compare this variable to
`NULL`.

* basic.c (gotoline):
* buffer.c (settabw):
* cscope.c (addentry):
* grep.c (compile_goto_error):
* kbd.c (setpromptdelay):
* main.c (main):
* paragraph.c (setfillcol): Do it.

+++ buffer.c
@@ -56,7 +56,7 @@ settabw(int f, int n)
 			return (ABORT);
 		if (bufp[0] == '\0')
 			return (ABORT);
-		width = strtonum(buf, INT_MIN, INT_MAX, &errstr);
+		width = (int) strtonum(buf, INT_MIN, INT_MAX, &errstr);
 		if (errstr != NULL) {
 			dobeep();
 			ewprintf("Invalid tab width: %s", bufp);
